Checkpoint work on simplifying Codec definition to use stack-based constructs whenever possible.
The previous approach relied on closures, which can be a pain to use in an unboxed fashion and
can lead to complicated movement of things to heap constructs.

In this approach, we turn Codec back into a trait, which means we lose some of the benefits
of composition while simplifying the definitions in other places.

More work is needed on the context codec operator and the hcodec! macros.