Convert more codecs to the new trait-based approach.
The codec functions that accept one or more codecs as an argument are in flux.  If we return
a Box<Codec<T>> we run into lifetime issues, but the workaround of returning an explicit
struct type is less than ideal, because we'd like to keep the codec impls private if possible.
Also, we're now running into issues with using the return value of said functions in a temporary
context.

One step forward, one or two steps back.