Box ALL the things.
The previous attempt used a mixture of static codecs and functions that return a boxed codec,
and things were getting hairy w.r.t. lifetime specifiers and keeping temporary things alive.

In this approach, we switch back to using functions everywhere (even for the integral codecs,
boo), but now everything takes and/or returns a Box<Codec<T>> so things are consistent.