Experiment with a new CodecRef sum type that allows for storing either an owned Box<Codec<T>> or a static ref.
In addition, we have an AsCodecRef<T> trait that allows functions like hlist_prepend_codec (i.e., that take one
or more codecs and create a new codec around them) to now accept either a boxed codec or a static one.

This takes advantage of a commonly used pattern in Rust, where traits are used to achieve something like
function overloading.

With all this in place, we can now switch back to using statics for codecs like the integral ones
that only need a single instance.  (The previous approach of using functions to return fresh boxed
codec instances wasn't the end of the world since these aren't very complex structures, but it did
seem to be fairly wasteful to create multiple instances when one shared instance would get the job
done.)